+++ src/main/java/com/invadermonky/thaumicapi/core/MixinLoader.java
@@ -4,8 +4,10 @@ import com.google.common.collect.ImmutableMap;
 import zone.rong.mixinbooter.Context;
 import zone.rong.mixinbooter.ILateMixinLoader;
 
-import java.util.*;
-import java.util.function.BooleanSupplier;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.function.Predicate;
 
 public class MixinLoader implements ILateMixinLoader {