+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## [1.1.0]
+### Added
+- Added NBTTagCompound support for Vis Discount items
+- Added vis discount support for items held in mainhand or offhand
+- Added helper method for adding Vis Discount to items `ThaumicAPI#addWarpingToStack(ItemStack, int)`
+- Added helper method for adding Warping to items `ThaumicAPI#addVisDiscountToStack(ItemStack, int)`
+
+---
+
 ## [1.0.2]
 ### Fixed
 - Fixed edge case Warp Event crash